pyproject.toml
[tool.poetry]
name = "redgen"
version = "0.1.0"
description = "A project that generates a video about a reddit post using tts."
authors = ["Nav <62173713+codeaye@users.noreply.github.com>"]
repository = "https://github.com/codeaye/redgen"
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
praw = "^7.7.0"
pyttsx3 = "^2.90"
requests = "^2.28.2"
moviepy = "^1.0.3"
selenium = "^4.8.3"
better-profanity = "^0.7.0"
pillow = "^9.5.0"
numpy = "^1.24.3"
beaupy = "^3.5.3"
setuptools = "^67.7.2"
proglog = "^0.1.10"
tqdm = "^4.65.0"
pytube = "^12.1.3"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.2"

[virtualenvs]
in-project = true

[tool.pyright]
exclude = [ ".venv" ]
venvPath = "."
venv = ".venv"

[tool.pylint.main]
disable = [
    "missing-function-docstring",
    "missing-module-docstring",
    "missing-class-docstring",
    "broad-exception-caught",
    "wrong-import-position",
    "unneeded-not"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

